Add tests for "Not Found" logging with a urlconf with i18n_patterns
These tests are very similar to the two tests from the previous
commit. The only difference is that language-prefixed URLs are
configured, (with i18n_patterns and LocaleMiddleware).

One of these tests (test_i18n_page_found_no_warning) is currently
expected to fail. This test demonstrates a bug that is currently
present: If a request comes in for a URL such as:

	/exists/

which LocaleMiddleware will redirect to:

	/en/exists/

Then a temporary 404 result is formed while middlewares are being
processed and this is logged as "Not Found: /exists/", before
LocaleMiddleware has a chance to fixup the 404 into a 302.

Ideally, no warning would be generated in the above case.